Fixed issue with drawing the wrong uv icon
- refactored the configs to use value objects instead of relying on
  validator
- removed lazy_static in favor of  once_cell
- switch from f64 to f32 and i32 to i16, cause these were not needed
- removed some unused template variables
- more documentation

+++ src/weather/utils.rs
@@ -29,7 +29,7 @@ pub fn get_moon_phase_icon_name() -> MoonPhaseIconName {
     let day = now.day();
 
     // Calculate the approximate age of the moon in days since the last new moon
-    let mut moon_age_days = ((year as f64 - 2000.0) * 365.25 + month as f64 * 30.6 + day as f64
+    let mut moon_age_days = ((year as f32 - 2000.0) * 365.25 + month as f32 * 30.6 + day as f32
         - 2451550.1)
         % 29.530588;
     if moon_age_days < 0.0 {
